Refactored Twitch code into it's own class
