+++ core/src/main/java/me/gosdev/chatpointsttv/TwitchEventHandler.java
@@ -25,11 +25,6 @@ public class TwitchEventHandler {
     Utils utils = ChatPointsTTV.getUtils();
 
     public static Boolean rewardBold;
-    private static ArrayList<String> lastRaids = new ArrayList<>();
-    private int lastRaidsIndex = 0;
-    Boolean listenForCheers = !plugin.config.getConfigurationSection("CHEER_REWARDS").getKeys(true).isEmpty();
-    Boolean listenForSubs = !plugin.config.getConfigurationSection("SUB_REWARDS").getKeys(true).isEmpty();
-    Boolean listenForGifts = !plugin.config.getConfigurationSection("GIFT_REWARDS").getKeys(true).isEmpty();
     Boolean logEvents = plugin.config.getBoolean("LOG_EVENTS");
     Boolean ignoreOfflineStreamers = plugin.config.getBoolean("IGNORE_OFFLINE_STREAMERS", false);
     ChatColor action_color = ChatPointsTTV.getChatColors().get("ACTION_COLOR").asBungee();
@@ -38,7 +33,7 @@ public class TwitchEventHandler {
     public void onChannelPointsRedemption(RewardRedeemedEvent event) {
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), event.getRedemption().getUser().getDisplayName() + " has redeemed " + event.getRedemption().getReward().getTitle() + " in " + TwitchUtils.getUsername(event.getRedemption().getChannelId()));
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getRedemption().getChannelId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
@@ -67,7 +62,7 @@ public class TwitchEventHandler {
     public void onFollow(ChannelFollowEvent event) {
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), event.getUserName() + " started following " + event.getBroadcasterUserName());
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getBroadcasterUserId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
@@ -95,7 +90,7 @@ public class TwitchEventHandler {
         if (event.getCheer() == null) return;
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), event.getChatterUserName() + " cheered " + event.getCheer().getBits() + " bits to " + event.getBroadcasterUserName() + "!");
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getBroadcasterUserId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
@@ -146,7 +141,7 @@ public class TwitchEventHandler {
 
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), event.getChatterUserName() + " has subscribed to " + event.getBroadcasterUserName() + " with a " + TwitchUtils.PlanToString(tier) + " sub!"); 
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getBroadcasterUserId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
@@ -179,7 +174,7 @@ public class TwitchEventHandler {
 
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), event.getChatterUserName() + " has gifted " + amount  + " " + tier + " subs in " + event.getBroadcasterUserName() + "'s' channel!"); 
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getBroadcasterUserId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
@@ -211,7 +206,7 @@ public class TwitchEventHandler {
 
         if (logEvents) utils.sendMessage(Bukkit.getConsoleSender(), raiderName + " has raided " + event.getToBroadcasterUserName()  + " with a viewer count of " + amount + "!"); 
         if (ignoreOfflineStreamers) {
-            for (Channel channel : plugin.getListenedChannels()) {
+            for (Channel channel : ChatPointsTTV.getTwitch().getListenedChannels()) {
                 if (channel.getChannelId().equals(event.getFromBroadcasterUserId()) && !channel.isLive()) return; // Return if channel matches and it's offline.
             }
         }
